Updating unit tests for pkg/client so that they pass with v1beta3 api
